django_server/requirements.txt
botocore==1.22.12
django-bootstrap-modal-forms==2.2.0
django-storages==1.12.3
whitenoise==5.3.0
psycopg2==2.9.1